+++ packages/spacecat-shared-http-utils/test/auth/handlers/ims.test.js
@@ -154,7 +154,7 @@ describe('AdobeImsHandler', () => {
     const result = await handler.checkAuth({}, testContext);
 
     expect(result).to.be.null;
-    expect(logStub.error.calledWith('[ims] Failed to validate token: Token not issued by expected idp: ims-na1-stg1 != ims-na1')).to.be.true;
+    expect(logStub.debug.calledWith('[ims] Failed to validate token: Token not issued by expected idp: ims-na1-stg1 != ims-na1')).to.be.true;
   });
 
   it('throw error when context is not correct', async () => {
@@ -178,7 +178,7 @@ describe('AdobeImsHandler', () => {
       const result = await handler.checkAuth({}, context);
 
       expect(result).to.be.null;
-      expect(logStub.error.calledWith('[ims] Failed to validate token: expires_in and created_at claims must be numbers')).to.be.true;
+      expect(logStub.debug.calledWith('[ims] Failed to validate token: expires_in and created_at claims must be numbers')).to.be.true;
     });
 
     it('returns null when expires_in is not a number', async () => {
@@ -188,7 +188,7 @@ describe('AdobeImsHandler', () => {
       const result = await handler.checkAuth({}, context);
 
       expect(result).to.be.null;
-      expect(logStub.error.calledWith('[ims] Failed to validate token: expires_in and created_at claims must be numbers')).to.be.true;
+      expect(logStub.debug.calledWith('[ims] Failed to validate token: expires_in and created_at claims must be numbers')).to.be.true;
     });
 
     it('returns null when created_at is in the future', async () => {
@@ -198,7 +198,7 @@ describe('AdobeImsHandler', () => {
       const result = await handler.checkAuth({}, context);
 
       expect(result).to.be.null;
-      expect(logStub.error.calledWith('[ims] Failed to validate token: created_at should be in the past')).to.be.true;
+      expect(logStub.debug.calledWith('[ims] Failed to validate token: created_at should be in the past')).to.be.true;
     });
 
     it('returns null when the token is expired', async () => {
@@ -208,7 +208,7 @@ describe('AdobeImsHandler', () => {
       const result = await handler.checkAuth({}, context);
 
       expect(result).to.be.null;
-      expect(logStub.error.calledWith('[ims] Failed to validate token: token expired')).to.be.true;
+      expect(logStub.debug.calledWith('[ims] Failed to validate token: token expired')).to.be.true;
     });
 
     it('successfully validates a token and returns the profile', async () => {